Support truthiness checks
This adds support for a `:key?` assertion syntax (a literal question mark
replacing the value) which checks for existence of the key and truthiness
of the value.

Also make sure that object methods are called only once to retrieve the
value in a test assertion, in case they have side effects.